use app standby to force stop background app

+++ src/me/piebridge/prevent/framework/ActivityReceiver.java
@@ -297,7 +297,9 @@ abstract class ActivityReceiver extends BroadcastReceiver {
                 break;
             }
             if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP || !hasHighPriority(packageName)) {
-                HideApiUtils.forceStopPackage(mContext, packageName);
+                if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M || !SystemHook.isUseAppStandby() || !SystemHook.inactive(packageName)) {
+                    HideApiUtils.forceStopPackage(mContext, packageName);
+                }
                 leavingPackages.remove(packageName);
                 checkLeavingNext.remove(packageName);
             }
